wd: Delete
I needed it to be able to spawn a new i3(1) window
with terminal in the same directory the currently
focused terminal is operating.  It's a big kludge
and it works actually only under some specific
conditions.  Now I switched to tmux(1), which has
a config option for exactly this task.  So there's
no need in this script anymore.