audstrip: Correct generating temporary output name
mktemp(1) does substitute X'es with unique value
only when they're appended to the name (e.g.
tem.XXX). That means that putting X'es as name
suffix (which I did) effectively results in X'es
not being substituted and being printed as they
are. To fix that, use mktemp(1) only to generate
a unique prefix and make up a tmp_name separately.

Fixes:	77c30d54f4695356c81cc5393efb85b3b6414152