r: Remove
After all, in-place replacement is a powerful and
responsible operation (plus it is not needed that
often) and we are better off treating each case
individually, combining grep(1), find(1) and sed(1)
(probably make a simple script for each case).

Moreover, I would expect wrapper script to be more
safe, thus, it would require a dry-run mode or 'print
all matches' mode, which would not be very trivial to
do, because we would have to parse sed(1) commands
in order to extract search-pattern and subject it to
grep(1) in order to find files with matches.

And, finally, using sed(1) itself is much more
flexible in terms of pattern match.